refactor: removing statsplugins and requires watchfiles for --reload
